通过命令行参数来控制 是否清空旧数据

+++ dataBase.py
@@ -13,9 +13,13 @@ client = MongoClient()
 db = client.xm94630
 coll = db.stocks
 
-#删除全部原有数据
-print('清空全部旧数据')
-result = coll.delete_many({})
+def clearOldDatabase(bool=False):
+    if bool:
+        #删除全部原有数据
+        print('清空全部旧数据')
+        result = coll.delete_many({})
+    else:
+        print('在旧数据上追加')
 
 #把抓取的一条数据，保存到 mongo 数据库
 def save(data={}):
